feat: Update Accessor class to include get_archetype_size()
The Accessor class in the `accessor.hpp` file has been updated to include a new member function `get_archetype_size()`. This function returns the size of the archetypes vector in the Accessor class. This change enhances the functionality of the Accessor class and provides access to the archetype size for querying purposes.